Day 9 Star 2 updates

+++ 2018/java/src/main/java/org/interannette/day9/DoubleLinkedNode.java
@@ -2,9 +2,11 @@ package org.interannette.day9;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class DoubleLinkedNode {
-    int value;
+    BigInteger value;
     DoubleLinkedNode clockwise;
     DoubleLinkedNode counterClockwise;
 }
